debugging kernel and middlware
