+++ public/index.php
@@ -2,10 +2,16 @@
 
 declare(strict_types=1);
 
+use App\Http\Middleware\HelloMiddleware;
 use App\Services\Http\Kernel;
 use Symfony\Component\HttpFoundation\Request;
 
 $container = require __DIR__ . '/../bootstrap/container.php';
 
-$requestHandler = (new Kernel())->process(Request::createFromGlobals());
+$middleware = [
+    HelloMiddleware::class,
+];
+
+$requestHandler = (new Kernel($middleware))->process(Request::createFromGlobals());
+
 $requestHandler->send();
