+++ src/Http/Middleware/MiddlewareInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Middleware;
+namespace App\Http\Middleware;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -11,7 +11,7 @@ interface MiddlewareInterface
 {
     /**
      * @param  \Symfony\Component\HttpFoundation\Request  $request
-     * @param  \App\Services\Middleware\StackInterface    $stack
+     * @param  \App\Http\Middleware\StackInterface        $stack
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
