+++ src/Services/Http/KernelInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Services\Http;
 
-use App\Services\Middleware\StackInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
