--- src/Services/Middleware/StackInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Services\Middleware;
-
-interface StackInterface
-{
-    public function next(): MiddlewareInterface;
-}
